+++ json/test/TSGeneratorTest.kt
@@ -4,6 +4,7 @@ import ch.tutteli.atrium.api.fluent.en_GB.toEqual
 import ch.tutteli.atrium.api.verbs.expect
 import org.junit.jupiter.api.Test
 import java.time.LocalDate
+import java.time.LocalTime
 import java.util.*
 import kotlin.reflect.KProperty1
 
@@ -27,8 +28,8 @@ class TSGeneratorTest {
       "interface Person {hello: SomeEnum; name: string}")
 
     expect(ts.render(SomeData::class)).toEqual( // language=TypeScript
-      "interface SomeData {age: number; any: any; birthDate?: string; bytes: Array<number>; field: keyof Person; id: MyId<SomeData>; " +
-        "list: Array<SomeData>; map: Record<string, Array<SomeData>>; name: string; other?: SomeData; status: SomeDataStatus; hello: SomeEnum}")
+      "interface SomeData {age: number; any: any; birthDate?: `$tsDate`; bytes: Array<number>; field: keyof Person; id: MyId<SomeData>; " +
+        "list: Array<SomeData>; map: Record<`$tsTime`, Array<SomeData>>; name: string; other?: SomeData; status: SomeDataStatus; hello: SomeEnum}")
   }
 }
 
@@ -40,10 +41,10 @@ class TSGeneratorTest {
 
 enum class SomeEnum { HELLO, WORLD }
 
-interface Person { val name: String; val hello get() = SomeEnum.HELLO }
+interface Person { val name: String; val hello get() = SomeEnum.HELLO; }
 
-data class SomeData(override val name: String, val age: Int, val birthDate: String?, val id: MyId<SomeData>, val other: SomeData?,
-                    val list: List<SomeData>, val map: Map<LocalDate, Array<SomeData>>, val any: Any, val status: Status = Status.ACTIVE,
+data class SomeData(override val name: String, val age: Int, val birthDate: LocalDate?, val id: MyId<SomeData>, val other: SomeData?,
+                    val list: List<SomeData>, val map: Map<LocalTime, Array<SomeData>>, val any: Any, val status: Status = Status.ACTIVE,
                     val field: KProperty1<Person, *>, val bytes: ByteArray): Person {
   enum class Status { ACTIVE }
 }